fix: update option name
