+++ src/PeraWalletConnect.ts
@@ -3,7 +3,6 @@ import WalletConnect from "@walletconnect/client";
 
 import PeraWalletConnectError from "./util/PeraWalletConnectError";
 import {
-  openPeraWalletConnectModal,
   openPeraWalletRedirectModal,
   removeModalWrapperFromDOM,
   PERA_WALLET_CONNECT_MODAL_ID,
@@ -14,9 +13,9 @@ import {
   closePeraWalletSignTxnModal,
   PERA_WALLET_IFRAME_ID,
   PERA_WALLET_MODAL_CLASSNAME,
-  PeraWalletModalConfig,
   PERA_WALLET_SIGN_TXN_MODAL_ID,
-  setupPeraWalletConnectModalCloseListener
+  setupPeraWalletConnectModalCloseListener,
+  generatePeraWalletConnectModalActions
 } from "./modal/peraWalletConnectModalUtils";
 import {
   getWalletDetailsFromStorage,
@@ -33,34 +32,14 @@ import {
   formatJsonRpcRequest
 } from "./util/transaction/transactionUtils";
 import {detectBrowser, isMobile} from "./util/device/deviceUtils";
-import {AlgorandChainIDs} from "./util/peraWalletTypes";
+import {
+  PeraWalletConnectMethodOptions,
+  PeraWalletConnectOptions
+} from "./util/peraWalletTypes";
 import {generateEmbeddedWalletURL} from "./util/peraWalletUtils";
 import appTellerManager, {PeraTeller} from "./util/network/teller/appTellerManager";
 import {getPeraWebWalletURL} from "./util/peraWalletConstants";
 import {getMetaInfo, waitForTabOpening} from "./util/dom/domUtils";
-
-interface PeraWalletConnectOptions {
-  bridge?: string;
-  shouldShowSignTxnToast?: boolean;
-  chainId?: AlgorandChainIDs;
-}
-
-function generatePeraWalletConnectModalActions({
-  isWebWalletAvailable,
-  shouldDisplayNewBadge,
-  shouldUseSound,
-  allowToSelectMultipleAccounts
-}: PeraWalletModalConfig) {
-  return {
-    open: openPeraWalletConnectModal({
-      isWebWalletAvailable,
-      shouldDisplayNewBadge,
-      shouldUseSound,
-      allowToSelectMultipleAccounts
-    }),
-    close: () => removeModalWrapperFromDOM(PERA_WALLET_CONNECT_MODAL_ID)
-  };
-}
 
 class PeraWalletConnect {
   bridge: string;
@@ -98,7 +77,8 @@ class PeraWalletConnect {
     resolve: (accounts: string[]) => void,
     reject: (reason?: any) => void,
     webWalletURL: string,
-    chainId: number | undefined
+    chainId: number | undefined,
+    shouldSelectSingleAccount: boolean
   ) {
     const browser = detectBrowser();
     const webWalletURLs = getPeraWebWalletURL(webWalletURL);
@@ -143,7 +123,8 @@ class PeraWalletConnect {
                   type: "CONNECT",
                   data: {
                     ...getMetaInfo(),
-                    chainId
+                    chainId,
+                    shouldSelectSingleAccount
                   }
                 },
 
@@ -243,7 +224,8 @@ class PeraWalletConnect {
               type: "CONNECT",
               data: {
                 ...getMetaInfo(),
-                chainId
+                chainId,
+                shouldSelectSingleAccount
               }
             },
 
@@ -264,7 +246,8 @@ class PeraWalletConnect {
                   type: "CONNECT",
                   data: {
                     ...getMetaInfo(),
-                    chainId
+                    chainId,
+                    shouldSelectSingleAccount
                   }
                 },
 
@@ -338,7 +321,7 @@ class PeraWalletConnect {
     };
   }
 
-  connect(allowToSelectMultipleAccounts?: boolean) {
+  connect({shouldSelectSingleAccount = false}: PeraWalletConnectMethodOptions = {}) {
     return new Promise<string[]>(async (resolve, reject) => {
       try {
         // check if already connected and kill session first before creating a new one.
@@ -363,7 +346,8 @@ class PeraWalletConnect {
           resolve,
           reject,
           webWalletURL,
-          this.chainId
+          this.chainId,
+          shouldSelectSingleAccount
         );
 
         if (isWebWalletAvailable) {
@@ -378,7 +362,7 @@ class PeraWalletConnect {
             isWebWalletAvailable,
             shouldDisplayNewBadge,
             shouldUseSound,
-            allowToSelectMultipleAccounts: allowToSelectMultipleAccounts || true
+            shouldSelectSingleAccount
           })
         });
