+++ src/modal/peraWalletConnectModalUtils.ts
@@ -7,7 +7,7 @@ export interface PeraWalletModalConfig {
   isWebWalletAvailable: boolean;
   shouldDisplayNewBadge: boolean;
   shouldUseSound: boolean;
-  allowToSelectMultipleAccounts: boolean;
+  shouldSelectSingleAccount: boolean;
 }
 
 // The ID of the wrapper element for PeraWalletConnectModal
@@ -42,11 +42,28 @@ function createModalWrapperOnDOM(modalId: string) {
   return wrapper;
 }
 
+function generatePeraWalletConnectModalActions({
+  isWebWalletAvailable,
+  shouldDisplayNewBadge,
+  shouldUseSound,
+  shouldSelectSingleAccount
+}: PeraWalletModalConfig) {
+  return {
+    open: openPeraWalletConnectModal({
+      isWebWalletAvailable,
+      shouldDisplayNewBadge,
+      shouldUseSound,
+      shouldSelectSingleAccount
+    }),
+    close: () => removeModalWrapperFromDOM(PERA_WALLET_CONNECT_MODAL_ID)
+  };
+}
+
 function openPeraWalletConnectModal(modalConfig: PeraWalletModalConfig) {
   return (uri: string) => {
     if (!document.getElementById(PERA_WALLET_CONNECT_MODAL_ID)) {
       const root = createModalWrapperOnDOM(PERA_WALLET_CONNECT_MODAL_ID);
-      const newURI = `${uri}&algorand=true&allowToSelectMultipleAccounts=${modalConfig.allowToSelectMultipleAccounts}`;
+      const newURI = `${uri}&algorand=true&allowToSelectMultipleAccounts=${modalConfig.shouldSelectSingleAccount}`;
       const {isWebWalletAvailable, shouldDisplayNewBadge, shouldUseSound} = modalConfig;
 
       root.innerHTML = `<pera-wallet-connect-modal uri="${newURI}" is-web-wallet-avaliable="${isWebWalletAvailable}" should-display-new-badge="${shouldDisplayNewBadge}" should-use-sound="${shouldUseSound}"></pera-wallet-connect-modal>`;
@@ -153,5 +170,6 @@ export {
   closePeraWalletSignTxnToast,
   removeModalWrapperFromDOM,
   openPeraWalletSignTxnModal,
-  closePeraWalletSignTxnModal
+  closePeraWalletSignTxnModal,
+  generatePeraWalletConnectModalActions
 };
