+++ src/util/network/teller/appTellerManager.ts
@@ -5,6 +5,7 @@ export type PeraTeller =
   | {
       type: "CONNECT";
       data: {
+        shouldSelectSingleAccount: boolean;
         title: string;
         url: string;
         favicon?: string;
